fix(ci): clear mypy, bandit, npm audit and the mcp-servers install
The pushed CI run failed six jobs. Reproducing each locally found real
bugs, not just lint noise.

mypy went 93 errors to zero. The genuine ones:

- The Anthropic provider passed a hand-rolled _NotGiven sentinel where the
  SDK checks isinstance(x, anthropic.NotGiven). A different class fails that
  check, so "omitted" parameters were being serialised into the request body
  instead. Any call with no tools or no stop sequences sent a garbage object
  to the API. Now uses the SDK's own NOT_GIVEN.
- Both the OpenAI and Anthropic stream() paths silently dropped tools, stop
  and response_schema — a streamed tool-calling turn could never invoke a
  tool, because the API was never told the tools existed. Both now share one
  kwargs builder with complete().
- The Gemini stream() path labelled every message "user", including the
  model's own prior turns, corrupting history on every streamed follow-up.
- GET /api/admin/prompts called sorted() on a bound method rather than its
  result: a guaranteed TypeError the first time anyone opened that page.
- gdpr.py referenced settings.opensearch_index and settings.neo4j_url,
  neither of which exists, and assumed one shared BM25 index when sparse.py
  uses per-tenant indices. Erasure now drops the tenant's own index.
- ingestion.py called graph_extraction.extract_and_store, which does not
  exist. Wrapped in a broad except, it had been silently no-opping.
- ExperimentOut was constructed without is_shadow or primary_metric, both
  required and both present on the model.
- A paragraph with no style crashed the DOCX parser on .name.

bandit: all ten findings already carried explanations — written as
`# noqa:`, which is ruff's syntax and which bandit does not read. Added the
`# nosec` markers the tool actually looks for.

mcp-servers had no [build-system] table, so `pip install -e` had nothing to
build against; that job was failing in eight seconds before pytest ran.
Verified in a clean venv: installs, 82 tests pass.

npm audit: patched postcss and vite via overrides rather than forcing a
breaking Next major bump. High and critical are clear; two low remain.

938 backend tests, 82 MCP tests, frontend build and typecheck all green.

+++ apps/api/src/caching/base.py
@@ -431,4 +431,4 @@ def cosine_similarity(a: Sequence[float], b: Sequence[float]) -> float:
     norm_b = sum(y * y for y in b) ** 0.5
     if norm_a == 0.0 or norm_b == 0.0:
         return 0.0
-    return dot / (norm_a * norm_b)
+    return float(dot / (norm_a * norm_b))
